workers/simulation/tasks/app.py
from celery import Celery
import os
from dotenv import load_dotenv

load_dotenv()

app = Celery(
    'portfolio-analytics',
    broker=os.getenv('CELERY_BROKER_URL', 'redis://localhost:6379/0'),
    backend=os.getenv('CELERY_RESULT_BACKEND', 'redis://localhost:6379/0'),
    include=['tasks.monte_carlo']
)

app.conf.update(
    task_serializer='json',
    accept_content=['json'],
    result_serializer='json',
    timezone='UTC',
    enable_utc=True,
    task_track_started=True,
    task_time_limit=600,  # 10 minutes max
    worker_prefetch_multiplier=1,
)

if __name__ == '__main__':
    app.start()
