workers/simulation/requirements.txt
celery[redis]==5.3.6
psycopg2-binary==2.9.9
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0
openai==1.12.0
anthropic==0.18.0
python-dotenv==1.0.1
redis==5.0.1
